refactor: quarantine engineer-harness verbs under cluster/
seictl carries two distinct surfaces: the node-operator's tool
(sidecar HTTP, config/genesis patching) and the engineer-facing
harness for harbor workloads (`bench`, `context`, future `onboard` /
`stress` / `integration`). The latter has a different audience, a
different contract surface (JSON envelopes, MCP-ready), and a different
lifecycle. Until the workload abstraction stabilizes (likely a generic
CRD wrapping benchmarks / stress / integration tests), the harness
hand-rolls each verb — and lives under `cluster/` so a future split
into its own binary is a `git mv`, not a multi-day untangle.

Move:
- bench.go, context.go, *_test.go         → cluster/
- internal/{clioutput,identity,validate,
            kube,aws,render}/             → cluster/internal/
- templates/                              → cluster/templates/

Wire:
- BenchCmd, ContextCmd are now exported from cluster/
- main.go imports cluster and registers them alongside the existing
  node-ops verbs

Docs:
- cluster/README.md — quarantine intent, layout, rules for new verbs
- CLAUDE.md — split the architecture section by surface

No functional change. All tests still pass; smoke against `bench up`
dry-run + `context` returns identical envelopes to slice 2/3a.
