+++ cluster/templates/embed.go
