+++ cluster/templates/snd.yaml
