+++ main.go
@@ -8,6 +8,8 @@ import (
 	"time"
 
 	"github.com/urfave/cli/v3"
+
+	"github.com/sei-protocol/seictl/cluster"
 )
 
 var (
@@ -49,8 +51,8 @@ var (
 			&awaitCmd,
 			&serveCmd,
 			&reportCmd,
-			&contextCmd,
-			&benchCmd,
+			&cluster.ContextCmd,
+			&cluster.BenchCmd,
 		},
 		Flags: []cli.Flag{
 			&cli.StringFlag{
